Move some functions to separate utils file
