+++ testing/tests/DevExpress.core/utils.dom.tests.js
@@ -6,8 +6,6 @@ import devices from "core/devices";
 import initMobileViewport from "mobile/init_mobile_viewport";
 import keyboardMock from "../../helpers/keyboardMock.js";
 
-const windowHeight = $(window).height();
-
 QUnit.module("createMarkup");
 
 QUnit.test("createMarkupFromString", function(assert) {
@@ -163,63 +161,4 @@ QUnit.test("it correctly detect the body element", function(assert) {
     var body = document.body;
 
     assert.ok(domUtils.contains(document, body), "Document contains the body element");
-});
-
-
-QUnit.module("style utils", {
-    beforeEach: function() {
-        this.$container = $("<div style='width: 100px; height: 100px; padding: 10px; box-sizing: border-box; margin: 5px'></div>").appendTo("#qunit-fixture");
-        this.$invisibleElement = $("<div style='width: 50px; height: 50px; display: none; padding: 5px;'></div>");
-        this.$container.append(this.$invisibleElement);
-    },
-
-    afterEach: function() {
-    }
-});
-
-QUnit.test("check calculateMaxHeight", function(assert) {
-    const checkFunc = ({ value, container, offset }, expected) => {
-        assert.strictEqual(domUtils.calculateMaxHeight(value, container, offset), expected);
-    };
-
-    checkFunc({ value: 300, container: null, offset: 0 }, 300);
-    checkFunc({ value: 300, container: null, offset: -100 }, 200);
-    checkFunc({ value: "300", container: null, offset: -100 }, 200);
-    checkFunc({ value: "300px", container: null, offset: -100 }, 200);
-    checkFunc({ value: "100mm", container: null, offset: -50 }, "calc(100mm - 50px)");
-    checkFunc({ value: "100pt", container: null, offset: -50 }, "calc(100pt - 50px)");
-    checkFunc({ value: "auto", container: null, offset: -50 }, "none");
-    checkFunc({ value: null, container: null, offset: -50 }, "none");
-
-    assert.roughEqual(domUtils.calculateMaxHeight("50%", window, -20), windowHeight / 2 - 20, 1, "string value in percent");
-});
-
-QUnit.test("check calculateMinHeight", function(assert) {
-    const checkFunc = ({ value, container, offset }, expected) => {
-        assert.strictEqual(domUtils.calculateMinHeight(value, container, offset), expected);
-    };
-
-    checkFunc({ value: 300, container: null, offset: 0 }, 300);
-    checkFunc({ value: 300, container: null, offset: -100 }, 200);
-    checkFunc({ value: "300", container: null, offset: -100 }, 200);
-    checkFunc({ value: "300px", container: null, offset: -100 }, 200);
-    checkFunc({ value: "100mm", container: null, offset: -50 }, "calc(100mm - 50px)");
-    checkFunc({ value: "100pt", container: null, offset: -50 }, "calc(100pt - 50px)");
-    checkFunc({ value: "auto", container: null, offset: -50 }, 0);
-    checkFunc({ value: null, container: null, offset: -50 }, 0);
-
-    assert.roughEqual(domUtils.calculateMinHeight("50%", window, -20), windowHeight / 2 - 20, 1, "string value in percent");
-});
-
-QUnit.test("check getPaddingsHeight", function(assert) {
-    assert.strictEqual(domUtils.getPaddingsHeight(null), 0, "no element");
-    assert.strictEqual(domUtils.getPaddingsHeight(this.$container), 20, "container paddings");
-    assert.strictEqual(domUtils.getPaddingsHeight(this.$container, true), 30, "include margins");
-    assert.strictEqual(domUtils.getPaddingsHeight(this.$invisibleElement), 10, "element paddings");
-});
-
-QUnit.test("check getVisibleHeight", function(assert) {
-    assert.strictEqual(domUtils.getPaddingsHeight(null), 0, "no element");
-    assert.strictEqual(domUtils.getVisibleHeight(this.$container), 100, "container height");
-    assert.strictEqual(domUtils.getVisibleHeight(this.$invisibleElement), 0, "invisible element height");
 });
